move distribution cli tests to e2e tests
