+++ tests/e2e/distribution/grpc_query_suite.go
@@ -1,4 +1,4 @@
-package testutil
+package distribution
 
 import (
 	"fmt"
