+++ tests/e2e/distribution/helpers.go
@@ -1,4 +1,4 @@
-package testutil
+package distribution
 
 import (
 	"bytes"
