+++ tests/e2e/distribution/withdraw_all_suite.go
@@ -1,4 +1,4 @@
-package testutil
+package distribution
 
 import (
 	"fmt"
